+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Add integration to collect Metrics, can be enabled by setting `options.enableMetrics = true` (#6956)
 
+### Fixes
+
+- Use `sendDefaultPii` and `span_id` for attributes in `BatcherScope` (#7055)
+
 ## 9.1.0
 
 > [!Warning]